remove azure functions from name and rename file
