+++ extensions/sql-database-projects/src/dialogs/addSqlBindingQuickpick.ts
